+++ mobile/android/android-components/components/support/ktx/src/main/java/mozilla/components/support/ktx/kotlin/String.kt
@@ -6,6 +6,9 @@ package mozilla.components.support.ktx.kotlin
 
 import android.net.Uri
 import android.text.TextUtils
+import java.text.SimpleDateFormat
+import java.util.Date
+import java.util.Locale
 
 /**
  * Normalizes a URL String.
@@ -36,3 +39,19 @@ fun String.isPhone(): Boolean = contains("tel:", true)
 fun String.isEmail(): Boolean = contains("mailto:", true)
 
 fun String.isGeoLocation(): Boolean = contains("geo:", true)
+
+/**
+ * Converts a [String] to a [Date] object.
+ * @param format date format used for formatting the this given [String] object.
+ * @param locale the locale to use when converting the String, defaults to [Locale.ROOT].
+ * @return a [Date] object with the values in the provided in this string, if empty string was provided, a current date
+ * will be returned.
+ */
+fun String.toDate(format: String, locale: Locale = Locale.ROOT): Date {
+    val formatter = SimpleDateFormat(format, locale)
+    return if (!this.isEmpty()) {
+        formatter.parse(this)
+    } else {
+        Date()
+    }
+}